# perf: transform ifnull into two conditions