--- a/frappe/tests/test_db_query.py
+++ b/frappe/tests/test_db_query.py
@@ -1200,6 +1200,14 @@ class TestDBQuery(IntegrationTestCase):
 		query = frappe.get_all("DocField", {"fieldname": None}, run=0)
 		self.assertIn("''", query)
 		self.assertNotIn("\\'", query)
+		self.assertNotIn("ifnull", query)
+
+	def test_ifnull_fallback_types(self):
+		query = frappe.get_all("DocField", {"fieldname": ("!=", None)}, run=0)
+		# Fallbacks should always be of correct type
+		self.assertIn("''", query)
+		self.assertNotIn("0", query)
+		self.assertNotIn("ifnull", query)
 
 
 class TestReportView(IntegrationTestCase):