`ifnull` just ensures that column index can't be used. Certain `ifnull` usage can be trivially broken into two separate conditions and both conditions can individually use indexes, so the query planner performs index intersection instead of full table scan.

This doesn't magically fix all problems, if underlying column isn't indexed then this change is of no use. BUT this at least gives a chance to optimizers to identify and auto tune queries... which is otherwise impossible.

Example:


```
MariaDB [_d90d7ebde9d08bc1]> analyze select name from `tabDocField` where fieldname = '';
+------+-------------+-------------+------+---------------+-----------+---------+-------+------+--------+----------+------------+--------------------------+
| id   | select_type | table       | type | possible_keys | key       | key_len | ref   | rows | r_rows | filtered | r_filtered | Extra                    |
+------+-------------+-------------+------+---------------+-----------+---------+-------+------+--------+----------+------------+--------------------------+
|    1 | SIMPLE      | tabDocField | ref  | fieldname     | fieldname | 563     | const | 1    | 0.00   |   100.00 |     100.00 | Using where; Using index |
+------+-------------+-------------+------+---------------+-----------+---------+-------+------+--------+----------+------------+--------------------------+
1 row in set (0.001 sec)

MariaDB [_d90d7ebde9d08bc1]> analyze select name from `tabDocField` where fieldname is null;
+------+-------------+-------------+------+---------------+-----------+---------+-------+------+--------+----------+------------+--------------------------+
| id   | select_type | table       | type | possible_keys | key       | key_len | ref   | rows | r_rows | filtered | r_filtered | Extra                    |
+------+-------------+-------------+------+---------------+-----------+---------+-------+------+--------+----------+------------+--------------------------+
|    1 | SIMPLE      | tabDocField | ref  | fieldname     | fieldname | 563     | const | 1    | 0.00   |   100.00 |     100.00 | Using where; Using index |
+------+-------------+-------------+------+---------------+-----------+---------+-------+------+--------+----------+------------+--------------------------+
1 row in set (0.001 sec)

MariaDB [_d90d7ebde9d08bc1]> analyze select name from `tabDocField` where fieldname is null or fieldname = '';
+------+-------------+-------------+-------------+---------------+-----------+---------+-------+------+--------+----------+------------+--------------------------+
| id   | select_type | table       | type        | possible_keys | key       | key_len | ref   | rows | r_rows | filtered | r_filtered | Extra                    |
+------+-------------+-------------+-------------+---------------+-----------+---------+-------+------+--------+----------+------------+--------------------------+
|    1 | SIMPLE      | tabDocField | ref_or_null | fieldname     | fieldname | 563     | const | 2    | 0.00   |   100.00 |     100.00 | Using where; Using index |
+------+-------------+-------------+-------------+---------------+-----------+---------+-------+------+--------+----------+------------+--------------------------+
1 row in set (0.010 sec)

MariaDB [_d90d7ebde9d08bc1]> analyze select name from `tabDocField` where coalesce(fieldname, '') = '';
+------+-------------+-------------+-------+---------------+-----------+---------+------+------+---------+----------+------------+--------------------------+
| id   | select_type | table       | type  | possible_keys | key       | key_len | ref  | rows | r_rows  | filtered | r_filtered | Extra                    |
+------+-------------+-------------+-------+---------------+-----------+---------+------+------+---------+----------+------------+--------------------------+
|    1 | SIMPLE      | tabDocField | index | NULL          | fieldname | 563     | NULL | 3058 | 3058.00 |   100.00 |       0.00 | Using where; Using index |
+------+-------------+-------------+-------+---------------+-----------+---------+------+------+---------+----------+------------+--------------------------+
1 row in set (0.011 sec)
```

TODO:
- [x] `=` operator
- [x] `!=` operator